Project Files/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cos-sdk==2.14.1
ibm-cos-sdk-core==2.14.1
ibm-cos-sdk-s3transfer==2.14.1
ibm_watsonx_ai==1.3.24
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
lomond==0.3.3
MarkupSafe==3.0.2
numpy==2.3.0
packaging==25.0
pandas==2.2.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tabulate==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
